# fe/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VAI Code",
    "description": "Weather Bridge AI production-oriented API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/subscriptions": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "List Subscriptions",
        "operationId": "list_subscriptions_api_v1_subscriptions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionResponse"
                  },
                  "type": "array",
                  "title": "Response List Subscriptions Api V1 Subscriptions Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Create Subscription",
        "operationId": "create_subscription_api_v1_subscriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subscriptions/consent": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "List Consents",
        "operationId": "list_consents_api_v1_subscriptions_consent_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsentResponse"
                  },
                  "type": "array",
                  "title": "Response List Consents Api V1 Subscriptions Consent Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Grant Consent",
        "operationId": "grant_consent_api_v1_subscriptions_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subscriptions/consent/{consent_id}": {
      "delete": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Withdraw Consent",
        "operationId": "withdraw_consent_api_v1_subscriptions_consent__consent_id__delete",
        "parameters": [
          {
            "name": "consent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Consent Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subscriptions/{subscription_id}": {
      "patch": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Update Subscription",
        "operationId": "update_subscription_api_v1_subscriptions__subscription_id__patch",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subscription Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/web-push/subscriptions": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Subscribe Web Push",
        "operationId": "subscribe_web_push_api_v1_notifications_web_push_subscriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebPushSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebPushSubscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/web-push/subscriptions/{contact_id}": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Unsubscribe Web Push",
        "operationId": "unsubscribe_web_push_api_v1_notifications_web_push_subscriptions__contact_id__delete",
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contact Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          },
          "withdrawn_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawn At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "purpose",
          "policy_version",
          "granted_at",
          "withdrawn_at"
        ],
        "title": "ConsentResponse"
      },
...
      "GrantConsentRequest": {
        "properties": {
          "policy_version": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Policy Version"
          }
        },
        "type": "object",
        "required": [
          "policy_version"
        ],
        "title": "GrantConsentRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubscriptionCreateRequest": {
        "properties": {
          "resident_location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resident Location Id"
          },
          "hazard_type": {
            "type": "string",
            "enum": [
              "flash_flood",
              "landslide",
              "fog"
            ],
            "title": "Hazard Type"
          },
          "minimum_level": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Minimum Level",
            "default": 1
          },
          "channel": {
            "type": "string",
            "enum": [
              "sms",
              "zalo",
              "email",
              "web_push",
              "webhook"
            ],
            "title": "Channel"
          },
          "quiet_hours_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours Start"
          },
          "quiet_hours_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours End"
          }
        },
        "type": "object",
        "required": [
          "hazard_type",
          "channel"
        ],
        "title": "SubscriptionCreateRequest"
      },
      "SubscriptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "resident_location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resident Location Id"
          },
          "hazard_type": {
            "type": "string",
            "title": "Hazard Type"
          },
          "minimum_level": {
            "type": "integer",
            "title": "Minimum Level"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "quiet_hours_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours Start"
          },
          "quiet_hours_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours End"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "resident_location_id",
          "hazard_type",
          "minimum_level",
          "channel",
          "quiet_hours_start",
          "quiet_hours_end",
          "is_active"
        ],
        "title": "SubscriptionResponse"
      },
      "SubscriptionUpdateRequest": {
        "properties": {
          "minimum_level": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Level"
          },
          "quiet_hours_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours Start"
          },
          "quiet_hours_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet Hours End"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "SubscriptionUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WebPushKeys": {
        "properties": {
          "p256dh": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "P256Dh"
          },
          "auth": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Auth"
          }
        },
        "type": "object",
        "required": [
          "p256dh",
          "auth"
        ],
        "title": "WebPushKeys"
      },
      "WebPushSubscriptionRequest": {
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Endpoint"
          },
          "expirationTime": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expirationtime"
          },
          "keys": {
            "$ref": "#/components/schemas/WebPushKeys"
          },
          "device_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Label"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "WebPushSubscriptionRequest"
      },
      "WebPushSubscriptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "is_active",
          "last_seen_at"
        ],
        "title": "WebPushSubscriptionResponse"
      }
    }
  }
}